Moved from controller to security file
